first block in epoch and core cleanup
